🔨 Move publish scripts into scripts folder
