+++ .changeset/three-cougars-yawn.md
@@ -1,0 +1,8 @@
+---
+'@codeshift/cli': patch
+'@codeshift/initializer': patch
+'@codeshift/publisher': patch
+'@codeshift/validator': patch
+---
+
+Update package configuration and refactors repo scripts
