+++ package.json
@@ -22,10 +22,11 @@
     "types:check": "tsc --noEmit --skipLibCheck",
     "monorepo:check": "manypkg check",
     "monorepo:fix": "manypkg fix && preconstruct fix",
-    "init:codemods": "ts-node scripts/initialize.ts",
     "start:codemods": "node packages/cli/bin/codeshift-cli.js",
+    "init:codemods": "ts-node scripts/initialize.ts",
     "validate:codemods": "ts-node scripts/validate.ts ./community",
-    "release:codemods": "ts-node packages/publisher/src/index.ts ./community ./.tmp",
+    "release:codemods": "ts-node scripts/publish.ts ./community ./.tmp",
+    "release-all:codemods": "ts-node scripts/publish-all.ts ./community ./.tmp",
     "prerelease": "yarn validate && yarn test",
     "release": "yarn changeset publish"
   },
